requirements.txt
# pip install -r requirements.txt 

langchain
langchain-openai
python-dotenv